Desolve the Finder, as it only introduces a circular dependency
