+++ lib/advisor/core/advice/advice.ex
@@ -1,6 +1,8 @@
 defmodule Advisor.Core.Advice do
   use Ecto.Schema
-  alias Advisor.Core.Advice.Finder
+  import Ecto.Query
+  alias Advisor.Repo
+  alias __MODULE__
 
   @primary_key {:id, :binary_id, autogenerate: true}
 
@@ -14,7 +16,15 @@ defmodule Advisor.Core.Advice do
     defstruct [:advisor, :advice_id]
   end
 
-  defdelegate find(id), to: Finder
-  defdelegate find(id, opts), to: Finder
-  defdelegate all_for(id), to: Finder
+  def all_for(id) do
+    Repo.all(from advice in Advice, where: advice.questionnaire_id == ^id)
+  end
+
+  def find(id) do
+    Repo.get(Advice, id)
+  end
+
+  def find(advice_id, [from_advisor: %{id: advisor_id}]) do
+    Repo.one(from a in Advice, where: a.id == ^advice_id and a.advisor_id == ^advisor_id)
+  end
 end
